requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
apache-airflow==2.7.0
sqlalchemy==1.4.49
